multiple(backend:repositories:answer)
test: added all + benchmark

fix: functions error return on empty slice

feat: two implementations of totals error returnment(are equal by bench)

+++ application/internal/models/usersGroupsTests.go
@@ -17,5 +17,5 @@ type UserGroupsTests struct {
 	UserUUID uuid.UUID `bun:"user_uuid,pk"`
 	User     *User     `bun:"rel:belongs-to,join:user_uuid=uuid"`
 
-	Score uint64 `bun:"score"`
+	Score int `bun:"score"`
 }
